refactored Loop.cpp to not calculate draw state multiple times, changed items list to simple vector, changed Id class to simple getId function

+++ src/modules/CollisionModule.cpp
@@ -49,11 +49,11 @@ void CollisionModule::update() {
 
     collisions.clear();
 
-    for (int i = 0; i < (int)bodies.size(); ++i) {
+    for (size_t i = 0; i < bodies.size(); ++i) {
 
         RigidBody body = bodies[i];
 
-        for (int j = 0; j < i; ++j) {
+        for (size_t j = 0; j < i; ++j) {
 
             RigidBody other = bodies[j];
